+++ src/main/java/com/google/cloud/tools/gradle/appengine/standard/StandardDeployTargetResolver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 Google Inc. All Right Reserved.
+ * Copyright 2018 Google LLC. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,55 +17,54 @@
 
 package com.google.cloud.tools.gradle.appengine.standard;
 
-import com.google.cloud.tools.appengine.AppEngineDescriptor;
-import com.google.cloud.tools.appengine.api.AppEngineException;
-import com.google.cloud.tools.gradle.appengine.core.DeployTargetResolver;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import org.gradle.api.GradleException;
-import org.xml.sax.SAXException;
+import static com.google.cloud.tools.gradle.appengine.core.ConfigReader.APPENGINE_CONFIG;
+import static com.google.cloud.tools.gradle.appengine.core.ConfigReader.GCLOUD_CONFIG;
 
-public class StandardDeployTargetResolver implements DeployTargetResolver {
+import com.google.cloud.tools.appengine.cloudsdk.Gcloud;
+import com.google.cloud.tools.gradle.appengine.core.ConfigReader;
+import java.io.File;
+import org.gradle.api.GradleException;
+
+public class StandardDeployTargetResolver {
 
   private final File appengineWebXml;
+  private final Gcloud gcloud;
 
-  public StandardDeployTargetResolver(File appengineWebXml) {
+  public StandardDeployTargetResolver(File appengineWebXml, Gcloud gcloud) {
     this.appengineWebXml = appengineWebXml;
+    this.gcloud = gcloud;
   }
 
-  @Override
+  /**
+   * Process user configuration of "projectId". If not configured, show usage. If set to
+   * APPENGINE_CONFIG then read from appengine-web.xml. If set to GCLOUD_CONFIG then read from
+   * gcloud's config state. If set but not a keyword then just return the set value.
+   */
   public String getProject(String configString) {
     if (configString == null || configString.trim().isEmpty()) {
       throw new GradleException(
-          "Deployment project must be defined or configured to read from system state\n"
-              + "1. Set appengine.deploy.project = 'my-project-name'\n"
-              + "2. Set appengine.deploy.project = '"
+          "Deployment projectId must be defined or configured to read from system state\n"
+              + "1. Set appengine.deploy.projectId = 'my-project-id'\n"
+              + "2. Set appengine.deploy.projectId = '"
               + APPENGINE_CONFIG
               + "' to use <application> from appengine-web.xml\n"
-              + "3. Set appengine.deploy.project = '"
+              + "3. Set appengine.deploy.projectId = '"
               + GCLOUD_CONFIG
               + "' to use project from gcloud config");
     } else if (configString.equals(APPENGINE_CONFIG)) {
-      try {
-        AppEngineDescriptor appEngineDescriptor =
-            AppEngineDescriptor.parse(new FileInputStream(appengineWebXml));
-        String appengineWebXmlProject = appEngineDescriptor.getProjectId();
-        if (appengineWebXmlProject == null || appengineWebXmlProject.trim().isEmpty()) {
-          throw new GradleException("<application> was not found in appengine-web.xml");
-        }
-        return appengineWebXmlProject;
-      } catch (IOException | SAXException | AppEngineException ex) {
-        throw new GradleException("Failed to read project from appengine-web.xml", ex);
-      }
+      return ConfigReader.getProject(appengineWebXml);
     } else if (configString.equals(GCLOUD_CONFIG)) {
-      return null;
+      return ConfigReader.getProject(gcloud);
     } else {
       return configString;
     }
   }
 
-  @Override
+  /**
+   * Process user configuration of "version". If not configured, show usage. If set to
+   * APPENGINE_CONFIG then read from appengine-web.xml. If set to GCLOUD_CONFIG then allow gcloud to
+   * generate a version. If set but not a keyword then just return the set value.
+   */
   public String getVersion(String configString) {
     if (configString == null || configString.trim().isEmpty()) {
       throw new GradleException(
@@ -78,18 +77,9 @@ public class StandardDeployTargetResolver implements DeployTargetResolver {
               + GCLOUD_CONFIG
               + "' to have gcloud generate a version for you.");
     } else if (configString.equals(APPENGINE_CONFIG)) {
-      try {
-        AppEngineDescriptor appEngineDescriptor =
-            AppEngineDescriptor.parse(new FileInputStream(appengineWebXml));
-        String appengineWebXmlVersion = appEngineDescriptor.getProjectVersion();
-        if (appengineWebXmlVersion == null || appengineWebXmlVersion.trim().isEmpty()) {
-          throw new GradleException("<version> was not found in appengine-web.xml");
-        }
-        return appengineWebXmlVersion;
-      } catch (IOException | SAXException | AppEngineException ex) {
-        throw new GradleException("Failed to read version from appengine-web.xml", ex);
-      }
+      return ConfigReader.getVersion(appengineWebXml);
     } else if (configString.equals(GCLOUD_CONFIG)) {
+      // can be null to allow gcloud to generate this
       return null;
     } else {
       return configString;